Move DatabaseManager to separate file
